Dream walker Added increasing/decreasing X and Y coordinates when left/right/up/down keys are pressed. Also added definition of initial position from world map file

+++ Engine/scenes/scene_world.h
@@ -69,8 +69,8 @@ public:
 
     QRect viewportRect;
 
-    long posX;
-    long posY;
+    double posX;
+    double posY;
     QString currentMusicFile;
 
     QString levelTitle;
